+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added an `UpdateStrategy` to `CollectionView` to allow specifying that it should update using non-
   animated `performBatchUpdates(…)`, which can be more performant and behave more predictably than
   `reloadData()`.
+- Added `reflowsForAccessibilityTypeSizes` and `forceVerticalAccessibilityLayout` properties to `HGroup.Style`.
 
 ### Fixed
 - Improve `CollectionView` logic for deciding when to `reloadData(…)` over `performBatchUpdates(…)`